Move NMEAParser, OdometryParser to lib/trackers.py
